DRTII-1472 Final tidying
