+++ server/src/main/scala/services/arrivals/RunnableMergedArrivals.scala
@@ -1,0 +1,77 @@
+package services.arrivals
+
+import actors.persistent.SortedActorRefSource
+import akka.NotUsed
+import akka.actor.ActorRef
+import akka.stream.scaladsl.{Flow, Sink}
+import akka.stream.{Materializer, UniqueKillSwitch}
+import akka.util.Timeout
+import drt.shared.CrunchApi.MillisSinceEpoch
+import providers.FlightsProvider
+import services.arrivals.MergeArrivals.FeedArrivalSet
+import services.crunch.deskrecs.QueuedRequestProcessing
+import uk.gov.homeoffice.drt.actor.commands.{MergeArrivalsRequest, ProcessingRequest}
+import uk.gov.homeoffice.drt.arrivals.{Arrival, ArrivalsDiff, UniqueArrival}
+import uk.gov.homeoffice.drt.ports.PortCode
+import uk.gov.homeoffice.drt.time.{DateLike, SDate, UtcDate}
+
+import scala.collection.SortedSet
+import scala.concurrent.{ExecutionContext, Future}
+
+object RunnableMergedArrivals {
+  def apply(portCode: PortCode,
+            flightsRouterActor: ActorRef,
+            aggregatedArrivalsActor: ActorRef,
+            mergeArrivalsQueueActor: ActorRef,
+            feedArrivalsForDate: Seq[DateLike => Future[FeedArrivalSet]],
+            mergeArrivalsQueue: SortedSet[ProcessingRequest],
+            mergeArrivalRequest: MillisSinceEpoch => MergeArrivalsRequest,
+            setPcpTimes: Seq[Arrival] => Future[Seq[Arrival]],
+            addArrivalPredictions: ArrivalsDiff => Future[ArrivalsDiff],
+           )
+           (implicit ec: ExecutionContext, mat: Materializer, timeout: Timeout): (ActorRef, UniqueKillSwitch) = {
+    val existingMergedArrivals: UtcDate => Future[Set[UniqueArrival]] =
+      (date: UtcDate) =>
+        FlightsProvider(flightsRouterActor)
+          .allTerminalsDateRange(date, date).map(_._2.map(_.unique).toSet)
+          .runWith(Sink.fold(Set[UniqueArrival]())(_ ++ _))
+          .map(_.filter(u => SDate(u.scheduled).toUtcDate == date))
+
+    val merger = MergeArrivals(
+      existingMergedArrivals,
+      feedArrivalsForDate,
+      ArrivalsAdjustments.adjustmentsForPort(portCode),
+    )
+
+    val mergeArrivalsFlow = MergeArrivals.processingRequestToArrivalsDiff(
+      mergeArrivalsForDate = merger,
+      setPcpTimes = setPcpTimes,
+      addArrivalPredictions = addArrivalPredictions,
+      updateAggregatedArrivals = aggregatedArrivalsActor ! _,
+    )
+
+    val (mergeArrivalsRequestQueueActor, mergeArrivalsKillSwitch: UniqueKillSwitch) =
+      startQueuedRequestProcessingGraph(
+        mergeArrivalsFlow,
+        mergeArrivalsQueueActor,
+        mergeArrivalsQueue,
+        flightsRouterActor,
+        "arrivals",
+        mergeArrivalRequest,
+      )
+    (mergeArrivalsRequestQueueActor, mergeArrivalsKillSwitch)
+  }
+
+  private def startQueuedRequestProcessingGraph[A](minutesProducer: Flow[ProcessingRequest, A, NotUsed],
+                                                   persistentQueueActor: ActorRef,
+                                                   initialQueue: SortedSet[ProcessingRequest],
+                                                   sinkActor: ActorRef,
+                                                   graphName: String,
+                                                   processingRequest: MillisSinceEpoch => ProcessingRequest,
+                                                  )
+                                                  (implicit materializer: Materializer): (ActorRef, UniqueKillSwitch) = {
+    val graphSource = new SortedActorRefSource(persistentQueueActor, processingRequest, initialQueue, graphName)
+    QueuedRequestProcessing.createGraph(graphSource, sinkActor, minutesProducer, graphName).run()
+  }
+
+}
